Fix visual glitch with bold/strikethrough.
I generated Gentium.fnt with Hiero, and apparently Hiero will use fully-transparent black even if the only color in the font is white (oddly, it still has fully-transparent white in the palette). This causes filtering issues if a font is stretched or shrunk and overdrawn, as was the case with bold, strikethrough, and sometimes underline. Fonts I made with Glamer don't seem to be affected.