Add AStarry font made with BMFont; needed some changes.
The width of a GlyphRegion was previously assumed to be close enough to its xAdvance, but that isn't true for all glyphs or all fonts. Now both xAdvance and width are used, and offsetX is also used differently. This seems to render both fixed-width and variable-width fonts better. Notably, 'f' renders fully in Gentium (instead of the round part getting clipped), while 'j' still can hook under a preceding glyph.