Big, huge refactor commit.
Big McLarge-Huge. This splits up TextraLayout into a series of Line objects, each with their own width and height but sharing a TextraFont. I'll probably get rid if the Textra- prefix... This allows multi-line text to work, and for markup to carry across lines if it was created from one markup() call.