Add TextraUITest, update TextraLabel and Font, among others.
TextraLabel needed some extra code to be given an appropriate size. This also involved Layout, a little. Font didn't always get the tiny snippet of a dash correctly for some fonts, and still probably doesn't... But now, it works with the default arial-15 font in libGDX, plus other fonts so far.